Added regester selection
user can use the first 5 switches (Left to right) of the DE10 FPGA board to select one of the 32 registers to view. The register valu eis then displayed on both the 7 segment displays and the LEDS. The most significant 24 bits are shown on the 6 7 segment displays. Then the remeaining 8 bits are displayed on the LEDs.